[MU4] Fix (MinGW/gcc) compiler warnings

+++ src/framework/mpe/tests/utils/articulationutils.h
@@ -40,8 +40,8 @@ inline PitchPattern createSimplePitchPattern(const pitch_level_t incrementDiff =
 {
     PitchPattern result;
 
-    for (int i = 0; i < EXPECTED_SIZE; ++i) {
-        result.pitchOffsetMap.insert_or_assign(i * TEN_PERCENTS, i * incrementDiff);
+    for (size_t i = 0; i < EXPECTED_SIZE; ++i) {
+        result.pitchOffsetMap.insert_or_assign(static_cast<int>(i) * TEN_PERCENTS, static_cast<int>(i) * incrementDiff);
     }
 
     return result;
@@ -53,8 +53,8 @@ inline ExpressionPattern createSimpleExpressionPattern(const dynamic_level_t amp
 
     float amplitudeSqrt = std::sqrt(amplitudeLevel);
 
-    for (int i = 0; i < EXPECTED_SIZE; ++i) {
-        duration_percentage_t currentPos = i * TEN_PERCENTS;
+    for (size_t i = 0; i < EXPECTED_SIZE; ++i) {
+        duration_percentage_t currentPos = static_cast<int>(i) * TEN_PERCENTS;
         dynamic_level_t value = amplitudeLevel - std::pow(
             (2 * (amplitudeSqrt / static_cast<float>(HUNDRED_PERCENTS)) * currentPos) - amplitudeSqrt, 2);
